refactored Products
